+++ src/Bitmap.h
@@ -1,5 +1,5 @@
 //A bitmap that stores image pixels as vectors
-//with floating-point color values (0-infinity)
+//with doubleing-point color values (0-infinity)
 //and maps them to 0..255 when exporting.
 #ifndef BITMAP_H
 #define BITMAP_H
@@ -9,7 +9,7 @@
 
 struct BitmapPixel{
     Vector color;
-    float depth;
+    double depth;
 };
 
 class Bitmap {
@@ -23,9 +23,9 @@ public:
     ~Bitmap();
 
     //Getters, setters
-    void setPixel(int x, int y, Vector pixel, float depth);
+    void setPixel(int x, int y, Vector pixel, double depth);
     Vector getPixel(int x, int y);
-    float getDepth(int x, int y);
+    double getDepth(int x, int y);
     int getWidth() {return width_;}
     int getHeight() {return height_;}
