+++ src/Plane.cpp
@@ -1,14 +1,14 @@
 #include "Plane.h"
 
 bool Plane::intersects(Ray ray) {
-    float denominator = ray.direction.dot(normal_);
-    float numerator = normal_.dot(point_ - ray.origin);
+    double denominator = ray.direction.dot(normal_);
+    double numerator = normal_.dot(point_ - ray.origin);
 
     //Numerator being zero could also mean that the ray lies completely in the
     //plane
     if (std::abs(denominator) < 0.0001) return false;
 
-    float distance = numerator / denominator;
+    double distance = numerator / denominator;
     if (distance < 0) return false;
     return true;
 }
@@ -16,8 +16,8 @@ bool Plane::intersects(Ray ray) {
 Intersection Plane::getIntersection(Ray ray) {
     Intersection result;
 
-    float denominator = ray.direction.dot(normal_);
-    float numerator = normal_.dot(point_ - ray.origin);
+    double denominator = ray.direction.dot(normal_);
+    double numerator = normal_.dot(point_ - ray.origin);
 
     if (std::abs(denominator) < 0.0001) {
         result.happened = false;
