+++ src/Vector.h
@@ -6,22 +6,22 @@
 
 class Vector {
 private:
-    float x_, y_, z_;
+    double x_, y_, z_;
 public:
     //Constructor
     Vector(){}
-    Vector(float x, float y, float z) {x_ = x; y_ = y; z_ = z;}
+    Vector(double x, double y, float z) {x_ = x; y_ = y; z_ = z;}
 
     //Getters
-    float getX() {return x_;}
-    float getY() {return y_;}
-    float getZ() {return z_;}
+    double getX() {return x_;}
+    double getY() {return y_;}
+    double getZ() {return z_;}
 
     //Setters
-    void setX(float x) {x_ = x;}
-    void setY(float y) {y_ = y;}
-    void setZ(float z) {z_ = z;}
-    void set(float x, float y, float z) {x_ = x; y_ = y; z_ = z;}
+    void setX(double x) {x_ = x;}
+    void setY(double y) {y_ = y;}
+    void setZ(double z) {z_ = z;}
+    void set(double x, double y, float z) {x_ = x; y_ = y; z_ = z;}
 
     //Algebra
     Vector& operator=(const Vector& rhs) {x_ = rhs.x_; y_ = rhs.y_; z_ = rhs.z_; return *this;}
@@ -29,22 +29,22 @@ public:
     //Compound operators
     Vector& operator+=(const Vector& rhs) {x_ += rhs.x_; y_ += rhs.y_; z_ += rhs.z_; return *this;}
     Vector& operator-=(const Vector& rhs) {x_ -= rhs.x_; y_ -= rhs.y_; z_ -= rhs.z_; return *this;}
-    Vector& operator/=(float rhs) {x_ /= rhs; y_ /= rhs; z_ /= rhs; return *this;}
-    Vector& operator*=(float rhs) {x_ *= rhs; y_ *= rhs; z_ *= rhs; return *this;}
+    Vector& operator/=(double rhs) {x_ /= rhs; y_ /= rhs; z_ /= rhs; return *this;}
+    Vector& operator*=(double rhs) {x_ *= rhs; y_ *= rhs; z_ *= rhs; return *this;}
 
     //Binary and unary operators
     Vector operator+(const Vector& rhs);
     Vector operator-(const Vector& rhs);
     Vector operator-();
 
-    Vector operator*(const float rhs) {return Vector(x_ * rhs, y_ * rhs, z_ * rhs);}
+    Vector operator*(const double rhs) {return Vector(x_ * rhs, y_ * rhs, z_ * rhs);}
 
     //Dot and cross products
-    float dot(const Vector& rhs) {return x_ * rhs.x_ + y_ * rhs.y_ + z_ * rhs.z_;}
+    double dot(const Vector& rhs) {return x_ * rhs.x_ + y_ * rhs.y_ + z_ * rhs.z_;}
     Vector cross(const Vector& rhs); //too long to be featured here
 
-    float normalize();
-    float modulus() {return sqrtf(x_ * x_ + y_ * y_ + z_ * z_);}
+    double normalize();
+    double modulus() {return sqrtf(x_ * x_ + y_ * y_ + z_ * z_);}
 
     //Reinhard mapping (0..inf to 0..1)
     void reinhardMap() {
@@ -53,7 +53,7 @@ public:
         z_ = z_ / (1 + z_);
     }
 
-    void toPow(float pow) {
+    void toPow(double pow) {
         x_ = powf(x_, pow);
         y_ = powf(y_, pow);
         z_ = powf(z_, pow);
